chore: ignore nullish coalescing for primitives

+++ apps/main/src/dex/components/PageCompensation/index.tsx
@@ -6,6 +6,7 @@ import { AlertFormError } from '@/dex/components/AlertFormError'
 import { Compensations } from '@/dex/components/PageCompensation/components/Compensations'
 import type { Balances, EtherContract, VestedTotals } from '@/dex/components/PageCompensation/types'
 import { CurveApi, ChainId, Provider } from '@/dex/types/main.types'
+import { Decimal } from '@primitives/decimal.utils'
 import { Box } from '@ui/Box'
 import { getErrorMessage } from '@ui-kit/utils'
 
@@ -56,12 +57,9 @@ export const FormCompensation = ({
       if (token === 'CRV') {
         try {
           const vestContract = new Contract(vestAddress, iface.format(), signer)
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment -- Existing violation before enabling this rule.
-          const unvested = await vestContract.lockedOf(contractAddress)
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment -- Existing violation before enabling this rule.
-          const fractions = await contracts[idx].contract.fractions(signerAddress)
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment -- Existing violation before enabling this rule.
-          const totalFraction = await contracts[idx].contract.total_fraction()
+          const unvested = (await vestContract.lockedOf(contractAddress)) as Decimal
+          const fractions = (await contracts[idx].contract.fractions(signerAddress)) as Decimal
+          const totalFraction = (await contracts[idx].contract.total_fraction()) as Decimal
 
           return { poolId, amount: ((Number(unvested) / 1e18) * Number(fractions)) / Number(totalFraction) }
         } catch (error) {
